Spliced Parser.h and Parser.cpp into a basic and a complex version
Having the entire parser library in one file can become cumbersome when working with several functions at a time. For this reason, I've decided to split the parser library into two versions based on a thought process.
If the function does not call another function in the parser library other than another instantation of itself, such as through recursion, it is a "basic function".
If the function requires several calls to different functions of the parser library, it is a "complex function".
This makes it easier to simply create a single-purpose function which can be put in the basic library, while functions which require more detail are put into the complex library.

In addition, the descriptions have been updated to provide a more concise walkthrough of the libraries.